requirements.txt
antlr4-python3-runtime==4.9.1
pillow
python-telegram-bot
